Extract helpers into module. NEXT: PointGroups::Update
