+++ app/mutations/point_groups/create.rb
@@ -1,6 +1,6 @@
 module PointGroups
   class Create < Mutations::Command
-    BAD_POINT_IDS = "The group contains invalid points."
+    include PointGroups::Helpers
 
     required do
       model :device, class: Device
@@ -23,18 +23,6 @@ module PointGroups
           pg.save!
           pg
         end
-      end
-    end
-
-    private
-
-    def points
-      @points ||= Point.where(id: point_ids, device: device)
-    end
-
-    def validate_point_ids
-      unless point_ids == points.pluck(:id).sort
-        add_error :points, :points_bad, BAD_POINT_IDS
       end
     end
   end
